# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Portfolio Dashboard API Documentation
  description: Complete API documentation for Portfolio Dashboard with Admin & Public authentication, OTP verification, and user management
  version: 2.0.0
  contact:
    name: Md Afzal Ansari
    email: mdafzal14777@gmail.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /admin-auth/register:
    post:
      summary: Register New Admin
      description: Register a new admin user with OTP verification
      tags:
        - Admin Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - user_name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "Admin User"
                  minLength: 2
                  maxLength: 50
                user_name:
                  type: string
                  example: "adminuser"
                  minLength: 3
                  maxLength: 20
                email:
                  type: string
                  format: email
                  example: "admin@example.com"
                password:
                  type: string
                  example: "password123"
                  minLength: 8
      responses:
        "200":
          description: Admin registration successful - OTP sent to email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Admin registration successful! Please check your email to verify your account."
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                      userType:
                        type: string
                        example: "admin"
                      expiresIn:
                        type: number
                        example: 600
        "400":
          description: Validation error
        "409":
          description: User already exists
...
  /admin-auth/registration-otp/status:
    get:
      summary: Check Admin Registration OTP Status
      description: Check if OTP exists and is active for admin registration
      tags:
        - Admin Authentication
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [registration]
      responses:
        "200":
          description: OTP status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      exists:
                        type: boolean
                      email:
                        type: string
                      type:
                        type: string
                      userType:
                        type: string
                      message:
                        type: string
  /admin-auth/registration-otp/resend:
    post:
      summary: Resend Admin Registration OTP
      description: Resend OTP for admin registration
      tags:
        - Admin Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - type
              properties:
                email:
                  type: string
                  format: email
                type:
                  type: string
                  enum: [registration]
      responses:
        "200":
          description: OTP resent successfully
...
  /admin-auth/otp/resend:
    post:
      summary: Resend OTP (Authenticated Admin)
      description: Resend OTP for various operations (password reset, email update, etc.)
      tags:
        - Admin OTP Operations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  enum: [password_reset, password_update, email_update, username_update]
                  example: "password_reset"
      responses:
        "200":
          description: OTP resent successfully
  /admin-auth/otp/status:
    get:
      summary: Check OTP Status (Authenticated Admin)
      description: Check status of active OTP for authenticated admin
      tags:
        - Admin OTP Operations
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [password_reset, password_update, email_update, username_update]
      responses:
        "200":
          description: OTP status retrieved successfully

  # ==================== PUBLIC USER AUTHENTICATION ====================
  /public-auth/register:
    post:
      summary: Register New Public User
      description: Register a new public user with OTP verification
      tags:
        - Public Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - user_name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "Public User"
                user_name:
                  type: string
                  example: "publicuser"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "password123"
      responses:
        "200":
          description: Public user registration successful - OTP sent to email
...
  /public-auth/registration-otp/status:
    get:
      summary: Check Public Registration OTP Status
      description: Check if OTP exists and is active for public user registration
      tags:
        - Public Authentication
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [registration]
      responses:
        "200":
          description: OTP status retrieved successfully
  /public-auth/registration-otp/resend:
    post:
      summary: Resend Public Registration OTP
      description: Resend OTP for public user registration
      tags:
        - Public Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - type
              properties:
                email:
                  type: string
                  format: email
                type:
                  type: string
                  enum: [registration]
      responses:
        "200":
          description: OTP resent successfully
...
  /public-auth/otp/resend:
    post:
      summary: Resend OTP (Authenticated Public User)
      description: Resend OTP for various operations (password reset, password update)
      tags:
        - Public OTP Operations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
                  enum: [password_reset, password_update]
                  example: "password_reset"
      responses:
        "200":
          description: OTP resent successfully
  /public-auth/otp/status:
    get:
      summary: Check OTP Status (Authenticated Public User)
      description: Check status of active OTP for authenticated public user
      tags:
        - Public OTP Operations
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [password_reset, password_update]
      responses:
        "200":
          description: OTP status retrieved successfully

# Security Schemes
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# Tags
...